refactor: split server error.rs tests into error/tests.rs
The metric_label mutation-coverage test in d5d1cd8 pushed error.rs to 563
lines, past the 500-line ratchet, and I pushed that commit anyway after
check_file_lengths.sh printed FAIL. Second time on this branch I have
committed past a gate I had just run — 3c27c14 was the first, on clippy.
Reading a gate's output is not optional, and neither is acting on it.

Fixed the way the codebase already solved this: a2a-protocol-client carries
`src/error/tests.rs`, split out when four stream-lag tests pushed *that* file
over the same limit. error.rs becomes error/mod.rs (212 lines) with the tests
in error/tests.rs (365), matching that precedent exactly.

The gate offers an exemption via `--update` as the alternative. Not taken:
the growth here is entirely test code, so splitting costs nothing and keeps
the ratchet meaningful, whereas recording an exemption would buy a larger
file and a weaker gate for no benefit. The gate's own wording — "makes it a
visible decision instead of a silent one" — cuts both ways, and this decision
did not need to be made at all.

fmt 0, clippy --workspace --all-targets --all-features 0, cargo test
--workspace --all-features 0 (82 "test result: ok"), error module 15/15,
check_file_lengths 0, check_proto_copies 0, check_mutation_scope 0,
mdbook build 0.
